+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## Unreleased
+## 0.6.0 - 2022-08-24
+### Added
+- `reconfigure`:
+  - Parameters can be now used within conditions.
+  - Implementation of `__contains__` to support `if ... in P`.
+  - `update()` now takes optional argument `only_existing` (def. `False`) to only update existing parameters.
+  - (ROS1 only) Parameters and their values are exposed to the ROS Parameter Server.
+  - `link(ConstrainedP, ConstrainedP)` to link two parameters together. First cannot be larger then the second one.
+
 ## 0.5.1 - 2022-03-23
 ### Added
 - `reconfigure`: